+++ utah-design-system-website/src/react/propTypesShapes/CharacterCountExamplePropsShape.js
@@ -1,0 +1,10 @@
+import PropTypes from 'prop-types';
+
+// uses CharacterCountExamplePropsShape
+
+export default PropTypes.exact({
+  className: PropTypes.string,
+  id: PropTypes.string,
+  maxLength: PropTypes.string,
+  text: PropTypes.string,
+});